fix: shrink tests
